% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eaMethods-package.R, R/zzz.R
\docType{package}
\name{eaMethods}
\alias{eaMethods}
\alias{eaMethods-package}
\title{Miscealleanous functions used for ecosystem assessments}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/david-beauchesne/eaMethods}
  \item Report bugs at \url{https://github.com/david-beauchesne/eaMethods/issues}
}

}
\author{
\strong{Maintainer}: David Beauchesne \email{david.beauchesne@hotmail.com} (\href{https://orcid.org/0000-0002-3590-8161}{ORCID}) [copyright holder]

}
\keyword{internal}
