% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lissage.R
\name{lissage}
\alias{lissage}
\title{Kernel weighted smoothing with arbitrary bounding area
Adapted from: https://www.r-bloggers.com/kernel-spatial-smoothing-transforming-points-pattern-to-continuous-coverage/}
\usage{
lissage(df, field, bandwidth, resolution, grid, out_crs = 32198)
}
\arguments{
\item{df}{sf object (points)}

\item{field}{weigth field in sf}

\item{resolution}{output grid resolution (map units)}

\item{grid}{sf study grid (polygon)}

\item{out_crs}{EPSG (should be an equal-area projection)}

\item{bandwith}{kernel bandwidth (map units)}
}
\value{
a raster object
}
\description{
Kernel weighted smoothing with arbitrary bounding area
Adapted from: https://www.r-bloggers.com/kernel-spatial-smoothing-transforming-points-pattern-to-continuous-coverage/
}
