% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_taxa.R
\name{review_taxa}
\alias{review_taxa}
\alias{clean_taxa}
\alias{remove_taxa}
\alias{combine_taxa}
\title{Function to clean, combine and remove taxa names}
\usage{
review_taxa(df, field, review = c("clean", "remove", "combine"))

clean_taxa(df, field)

remove_taxa(df, field)

combine_taxa(df, field)
}
\arguments{
\item{df}{data.frame containing the list of species to be cleaned}

\item{field}{name of the field containing the list of scientific species names}

\item{review}{what type(s) of review to perform between c("clean","remove","combine"). "clean" removes a certain number of text usually accompanying species names, such as "sp." and "spp.". "combine" modifies the names of species to combine certain species based on a list of species that should be combined (view code), and "remove" removes all species that are contained in a list of species that should be removed.}
}
\value{
This function returns the provided data.frame with the names of the species reviewd,
cleaned, combined and/or removed. Note that this function was initially built for the species
of the estuary and gulf of St. Lawrence and extended to include species of the North West
Atlantic. Using this function on any list should be done with care and be thhoroughly reviewed
afterwarsd. Also note that the justification for combinations and removals are documented in
the code.
}
\description{
This function clean taxa names, combines taxa that are similar taxonomically and that may
be hard to distinguish in the field, and removes taxa whose identification is too coarse to
be useful.
}
\examples{
df <- data.frame(
  species = c(
    "Boltenia c.,large", "Boltenia sp. (UNID)",
    "Triglops spp.", "Triglops murrayi",
    "Amphipoda"
  ),
  frequency = c(61, 100, 27, 50, 2)
)
field <- "species"

review_taxa(df, field, "clean")
review_taxa(df, field, c("clean", "remove"))
review_taxa(df, field, c("clean", "combine"))

}
\keyword{species}
\keyword{taxonomy}
