% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxa_clean}
\alias{taxa_clean}
\title{Strings to remove from the scientific names of species}
\format{
A data frame with 1 variable:
\describe{
\item{remove}{regex strings to remove from species names}
...
}
}
\usage{
taxa_clean
}
\description{
A dataset containing a list of regex strings used to clean up the scientific names of species
}
\keyword{datasets}
