% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxa_combine}
\alias{taxa_combine}
\title{Species to combine in a dataset}
\format{
A data frame with 3 variables:
\describe{
\item{to}{new taxonomic name to use}
\item{from}{taxonomic name to transform}
\item{documentation}{reasoning for the removal}
...
}
}
\usage{
taxa_combine
}
\description{
A dataset containing a list of taxonomic names to combine in a data.frame
}
\keyword{datasets}
