% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxa_remove}
\alias{taxa_remove}
\title{Species to remove from a dataset}
\format{
A data frame with 2 variable:
\describe{
\item{remove}{list of taxonomic names to remove from a data.frame}
\item{documentation}{reasoning for the removal}
...
}
}
\usage{
taxa_remove
}
\description{
A dataset containing a list of taxonomic  names to remove from a data.frame
}
\keyword{datasets}
