% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishing_intensity.R
\name{fishing_intensity}
\alias{fishing_intensity}
\title{Function to evaluate fishing intensity}
\usage{
fishing_intensity(fishData, areaGrid, metric = 1)
}
\arguments{
\item{fishData}{fishing data as sf object}

\item{areaGrid}{regular grid as sf object}

\item{metric}{type(s) of fishing intensity metric to evaluate: 1: Fishing effort density, no normalization; 2: Fishing effort density; 3: Fishing biomass yield density; 4: Fishing relative biomass yield density}
}
\description{
The function evaluates fishing intensity over a regular grid
}
\details{
This function uses fisheries data to evaluate the intensity of fishing over a regular grid
}
\keyword{fishing}
\keyword{intensity}
